新增参数校验

+++ src/main/java/com/fhpt/imageqmind/service/LoginService.java
@@ -14,7 +14,7 @@ public interface LoginService {
 
     boolean isLogin();
 
-    void login(Long userId);
+    void login(int userId);
 
-    void logout(Long userId);
+    void logout(int userId);
 }
